+++ verify/__meta__.py
@@ -8,7 +8,7 @@ __url__ = 'https://github.com/dgilland/verify'
 
 __version__ = '1.1.0'
 
-__install_requires__ = ['pydash>=3.2.2']
+__install_requires__ = ['pydash>=4.0.3']
 
 __author__ = 'Derrick Gilland'
 __email__ = 'dgilland@gmail.com'